[Feature] Better development environment
This commit includes configuration for Docker to run the project using
only one command: 'docker-compose up'. Gone are the days of dealing with
a local install of Apache or, god forbid, XAMPP. A development htaccess
file has also been created which will only be used during the
development process and is left out of the source entirely for
production.