Savedata is not an application directory
Savedata is not an application directory.
I have asked where the app is stored and was informed that we should mention where the configuration data is instead. It appears the Mac jargon is Application Support.
So update Download.php to be more accurate in the information provided.
Suggestions are welcome.